requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
yfinance==0.2.36
cachetools==5.3.2
python-dotenv==1.0.0
Werkzeug==3.0.1
tavily-python>=0.7.0
crawl4ai>=0.3.0
gunicorn==21.2.0
